requirements.txt
fastapi
uvicorn
langchain
langgraph
openai
python-dotenv
pydantic[email]
requests
smtplib
email-validator
